Added Dafny programs from lectures and tutorials of the 2023 iteration of the course
In the missing weeks (6, 9, 10, and 13), only powerpoint presentations + the white board were used
in the classroom, such that no new material was formulated in Dafny.

Note: in that iteration of the course we used at first Dafny 3.13.1, and later in the semester we
switched to Dafny 4.1.0. Ahead of adding the programs here, I've updated them to Dafny 4.11.0.